refactor: Go workspace — isolate bbolt in ext/boltdb + fix port conflict
Go workspace (go.work):
- Move boltdb materializer to ext/boltdb/ with its own go.mod
- Root go.mod no longer has go.etcd.io/bbolt dependency
- Dependabot/security scanners only see root go.mod
- ext/boltdb builds independently, used by venturi

Port conflict fix:
- When --control is set (spawned by daemon), auto-assign port
  if --http wasn't explicitly passed (localhost:0)
- Prevents "address already in use" when port 7532 is taken

Also: remove -tags boltdb from Taskfile (no longer in root module)
