+++ Taskfile.yml
@@ -32,8 +32,7 @@ tasks:
         sh: git describe --tags --abbrev=0 2>/dev/null || echo "dev"
     cmds:
       - >-
-        go build -tags boltdb
-        -ldflags "-X github.com/agentic-research/mache/cmd.Version={{.GIT_TAG}}
+        go build         -ldflags "-X github.com/agentic-research/mache/cmd.Version={{.GIT_TAG}}
         -X github.com/agentic-research/mache/cmd.Commit={{.GIT_COMMIT}}
         -X github.com/agentic-research/mache/cmd.Date={{.BUILD_DATE}}"
         -o {{.BINARY_NAME}} .
