+++ ext/boltdb/go.mod
@@ -1,0 +1,20 @@
+module github.com/agentic-research/mache/ext/boltdb
+
+go 1.25.0
+
+require (
+	go.etcd.io/bbolt v1.4.3
+	modernc.org/sqlite v1.49.1
+)
+
+require (
+	github.com/dustin/go-humanize v1.0.1 // indirect
+	github.com/google/uuid v1.6.0 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/ncruces/go-strftime v1.0.0 // indirect
+	github.com/remyoudompheng/bigfft v0.0.0-20230129092748-24d4a6f8daec // indirect
+	golang.org/x/sys v0.42.0 // indirect
+	modernc.org/libc v1.72.0 // indirect
+	modernc.org/mathutil v1.7.1 // indirect
+	modernc.org/memory v1.11.0 // indirect
+)
