+++ go.mod
@@ -13,7 +13,6 @@ require (
 	github.com/spf13/cobra v1.10.2
 	github.com/stretchr/testify v1.11.1
 	github.com/willscott/go-nfs v0.0.4
-	go.etcd.io/bbolt v1.4.3
 	golang.org/x/sys v0.43.0
 	golang.org/x/text v0.36.0
 	modernc.org/sqlite v1.49.1
